Projetos/Harven.ai 2/backend/requirements.txt
# ============================================
# HARVEN.AI - DEPENDÊNCIAS BACKEND
# ============================================

# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database & Storage
supabase>=2.3.0

# Environment & Configuration
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.1

# File Upload
python-multipart>=0.0.9

# PDF Text Extraction
pdfplumber>=0.10.0

# AI & OpenAI
openai>=1.12.0

# Security
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0

# HTTP Client (para healthchecks)
httpx>=0.26.0

# Utilities
python-dateutil>=2.8.2
